requirements.txt
certifi==2021.5.30
fastavro==1.4.1
pulsar-client==2.8.0
six==1.16.0
retry==0.9.2
viaa-chassis==0.2.0rc1
meemoo-cloudevents==0.1.0rc2